--- drive_train/launch/readme.txt
@@ -1,1 +1,0 @@
-don't use this file
