--- drive_train/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/drive_train
-[install]
-install_scripts=$base/lib/drive_train
