--- drive_train/urdf/basic_rover.urdf
@@ -1,74 +1,0 @@
-<?xml version="1.0"?>
-<robot name="basic_rover">
-    <link name="base_link">
-        <visual>
-            <geometry>
-                <box size="0.6 0.4 0.05"/>
-            </geometry>
-        </visual>
-    </link>
-    
-
-    <link name="right_front_wheel">
-        <visual>
-            <origin rpy="1.57075 0 0" xyz="0 0 0"/>            
-            <geometry>
-                <cylinder length="0.1" radius="0.075"/>
-            </geometry>
-        </visual>
-    </link>
-    <joint name="right_front_wheel_joint" type="continuous">
-        <axis rpy="0 0 0" xyz="0 1 0"/>
-        <parent link="base_link"/>
-        <child link="right_front_wheel"/>
-        <origin rpy="0 0 0" xyz="0.2625 -0.25 0"/>    
-    </joint>
-
-
-    <link name="left_front_wheel">
-        <visual>
-            <origin rpy="1.57075 0 0" xyz="0 0 0"/>            
-            <geometry>
-                <cylinder length="0.1" radius="0.075"/>
-            </geometry>
-        </visual>
-    </link>
-    <joint name="left_front_wheel_joint" type="continuous">
-        <axis rpy="0 0 0" xyz="0 1 0"/>
-        <parent link="base_link"/>
-        <child link="left_front_wheel"/>
-        <origin rpy="0 0 0" xyz="0.2625 0.25 0"/>    
-    </joint>
-
-
-    <link name="right_back_wheel">
-        <visual>
-            <origin rpy="1.57075 0 0" xyz="0 0 0"/>            
-            <geometry>
-                <cylinder length="0.1" radius="0.075"/>
-            </geometry>
-        </visual>
-    </link>
-    <joint name="right_back_wheel_joint" type="continuous">
-        <axis rpy="0 0 0" xyz="0 1 0"/>
-        <parent link="base_link"/>
-        <child link="right_back_wheel"/>
-        <origin rpy="0 0 0" xyz="-0.2625 0.25 0"/>    
-    </joint>
-
-
-    <link name="left_back_wheel">
-        <visual>
-            <origin rpy="1.57075 0 0" xyz="0 0 0"/>            
-            <geometry>
-                <cylinder length="0.1" radius="0.075"/>
-            </geometry>
-        </visual>
-    </link>
-    <joint name="left_back_wheel_joint" type="continuous">
-        <axis rpy="0 0 0" xyz="0 1 0"/>
-        <parent link="base_link"/>
-        <child link="left_back_wheel"/>
-        <origin rpy="0 0 0" xyz="-0.2625 -0.25 0"/>    
-    </joint>
-</robot>
