--- realsense_test.py
@@ -1,41 +1,0 @@
-import pyrealsense2 as rs
-import numpy as np
-import cv2
-
-# Configure depth and color streams
-pipeline = rs.pipeline()
-config = rs.config()
-config.enable_stream(rs.stream.depth, 640, 480, rs.format.z16, 30)
-config.enable_stream(rs.stream.color, 640, 480, rs.format.bgr8, 30)
-
-# Start the camera
-pipeline.start(config)
-
-try:
-    while True:#Breadcrumbs
-
-    #rover/rover_interfaces/msg
-        # Wait for a coherent pair of frames: depth and color
-        frames = pipeline.wait_for_frames()
-        depth_frame = frames.get_depth_frame()
-        color_frame = frames.get_color_frame()
-        if not depth_frame or not color_frame:
-            continue
-
-        # Convert images to numpy arrays
-        depth_image = np.asanyarray(depth_frame.get_data())
-        color_image = np.asanyarray(color_frame.get_data())
-
-        # Apply colormap on depth image (image must be converted to 8-bit per pixel first)
-        depth_colormap = cv2.applyColorMap(cv2.convertScaleAbs(depth_image, alpha=0.03), cv2.COLORMAP_JET)
-
-        # Display images
-        cv2.imshow('RealSense', np.hstack((color_image, depth_colormap)))
-
-        # Break loop with 'q' key
-        if cv2.waitKey(1) & 0xFF == ord('q'):
-            break
-finally:
-    # Stop streaming
-    pipeline.stop()
-    cv2.destroyAllWindows()
