--- rover_interfaces/.vscode/settings.json
@@ -1,12 +1,0 @@
-{
-    "python.autoComplete.extraPaths": [
-        "/home/ndt/Desktop/Terraformers-Rover/dev_ros2_ws/install/drive_train/lib/python3.10/site-packages",
-        "/opt/ros/humble/lib/python3.10/site-packages",
-        "/opt/ros/humble/local/lib/python3.10/dist-packages"
-    ],
-    "python.analysis.extraPaths": [
-        "/home/ndt/Desktop/Terraformers-Rover/dev_ros2_ws/install/drive_train/lib/python3.10/site-packages",
-        "/opt/ros/humble/lib/python3.10/site-packages",
-        "/opt/ros/humble/local/lib/python3.10/dist-packages"
-    ]
-}
