--- rover_pkg/CMakeLists.txt
@@ -1,51 +1,0 @@
-cmake_minimum_required(VERSION 3.8)
-project(rover_pkg)
-
-if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
-  add_compile_options(-Wall -Wextra -Wpedantic)
-endif()
-
-# find dependencies
-find_package(ament_cmake REQUIRED)
-#find_package(iostream REQUIRED)
-#find_package(fcntl REQUIRED)
-#find_package(termios REQUIRED)
-#find_package(unistd REQUIRED)
-#find_package(cstring REQUIRED)
-find_package(rclcpp REQUIRED)
-find_package(geometry_msgs REQUIRED)
-find_package(std_msgs REQUIRED)
-find_package(rover_interfaces REQUIRED)
-
-add_executable(driver_node_v2 src/driver_node_v2.cpp)
-ament_target_dependencies(
-  driver_node_v2                          
-  #iostream
-  #fcntl
-  #termios
-  #unistd
-  #cstring
-  rclcpp
-  geometry_msgs
-  std_msgs
-  rover_interfaces
-)
-
-install(TARGETS
-  driver_node_v2
-  DESTINATION lib/${PROJECT_NAME}
-)
-
-if(BUILD_TESTING)
-  find_package(ament_lint_auto REQUIRED)
-  # the following line skips the linter which checks for copyrights
-  # comment the line when a copyright and license is added to all source files
-  set(ament_cmake_copyright_FOUND TRUE)
-  # the following line skips cpplint (only works in a git repo)
-  # comment the line when this package is in a git repo and when
-  # a copyright and license is added to all source files
-  set(ament_cmake_cpplint_FOUND TRUE)
-  ament_lint_auto_find_test_dependencies()
-endif()
-
-ament_package()
