--- rover_pkg/package.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>rover_pkg</name>
-  <version>0.0.0</version>
-  <description>This packages contains code 
-                that will be run on the rover's
-                on-board computer.</description>
-  <maintainer email="thoennesn@gmail.com">ndt</maintainer>
-  <license>TODO: License declaration</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
