+++ src/comms_interfaces/msg/StereoCameraInfo.msg
@@ -1,0 +1,2 @@
+sensor_msgs/Image rgb
+sensor_msgs/Image depth
