--- src/rover_interfaces/msg/OdomInfo.msg
@@ -1,4 +1,0 @@
-float64 fl
-float64 fr
-float64 bl
-float64 br
