+++ src/rover_sensors/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/rover_sensors
+[install]
+install_scripts=$base/lib/rover_sensors
