--- src/station/package.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>station</name>
-  <version>0.0.0</version>
-  <description>TODO: Package description</description>
-  <maintainer email="srimans@umd.edu">sriman</maintainer>
-  <license>Apache-2.0</license>
-
-  <test_depend>ament_copyright</test_depend>
-  <test_depend>ament_flake8</test_depend>
-  <test_depend>ament_pep257</test_depend>
-  <test_depend>python3-pytest</test_depend>
-
-  <exec_depend>geometry_msgs</exec_depend>
-  <exec_depend>sensor_msgs</exec_depend>
-  <depend>rover_interfaces</depend>
-
-  <export>
-    <build_type>ament_python</build_type>
-  </export>
-</package>
