--- src/station/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/station
-[install]
-install_scripts=$base/lib/station
