+++ src/station_ui/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/station_ui
+[install]
+install_scripts=$base/lib/station_ui
